Update 20240402 - Began working on Discussion
Added related files for the Discussion (Forum) Module. This includes the migration, factory, seeder, model, and controller. The landing page now also contains the trending discussions, which consists of the top 4 categories with lots of discussions.

Furthermore, a bug fix is implemented for the Layout service provided where it throws and error if the initial migration failed and you tried to migrate again. This is due to the fact that the table is present but no content, which then results in an error when fetching some values.