Update 20240420 - Implemented Acc Reactivation
Implemented the Account Reactivation part with verification code request. Furthermore, confirm password middleware is now moved to the entire deactivate page along with the addition of new email layout templates and the contact-us.

Deactivated accounts will not show up when searched nor queried in discussion or anything. However, their data like the comments and discussions they started, will be retained and can be accessed by everyone on the website.