Update 20240421 - Implemented Reporting
Added the Report Feature, allowing users to report discussions and comments, which will then be saved into the database and allows admins and teachers to see. Reporter will remain anonymous to these people and will only be seen once the admin or teacher confirms their password.

This action will also log the said activity, allowing admins to determine who kept intruding over someone's privacy. (This part is to be done in the following commits)